# openapi.yaml
openapi: 3.1.0
info:
  title: AV Suite Backend Foundation
  description: Rehabilitation exercises management backend platform
  version: 0.1.0
paths:
...
  /api/v1/exercises:
    get:
      tags:
      - Exercises
      - Exercises
      summary: List Exercises
      description: "Endpoint ka purpose: Exercises list karna advanced filtering aur\
        \ pagination ke saath\nYeh endpoint exercise listing handle karta hai jismein:\n\
        1. Clinic-specific + global exercises fetch karte hain\n2. Multiple filters\
        \ apply karte hain (body_part, is_free, search)\n3. Pagination apply karte\
        \ hain\n4. Metadata ke saath results return karte hain\n\nHTTP Method: GET\n\
        URL: /api/v1/exercises\nStatus Code: 200 OK\n\nQuery Parameters:\n- body_part\
        \ (str, optional): Filter exercises by body part (e.g., \"Shoulder\", \"Knee\"\
        )\n- is_free (bool, optional): Filter by pricing (true = free, false = paid)\n\
        - search (str, optional): Search in exercise title (case-insensitive substring)\n\
        - page (int, default: 1): Page number for pagination\n- page_size (int, default:\
        \ 10): Results per page (max: 100)\n\nRequest Headers:\nAuthorization: Bearer\
        \ {jwt_token}\n\nResponse (ResponseEnvelope[List[ExerciseRead]]):\n{\n   \
        \ \"status\": \"success\",\n    \"data\": [\n        {\n            \"id\"\
        : \"uuid\",\n            \"clinic_id\": \"uuid or null\",\n            \"\
        title\": \"Shoulder Rotation\",\n            \"description\": \"Detailed description\"\
        ,\n            \"body_part\": \"Shoulder\",\n            \"is_free\": true,\n\
        \            \"video_url\": \"https://cdn.example.com/exercise.mp4\",\n  \
        \          \"created_at\": \"2024-01-01T00:00:00Z\",\n            \"updated_at\"\
        : \"2024-01-01T00:00:00Z\"\n        },\n        ...\n    ],\n    \"meta\"\
        : {\n        \"total\": 23,\n        \"page\": 1,\n        \"page_size\":\
        \ 10\n    }\n}\n\nError Responses:\n- 401 UNAUTHORIZED: Missing or invalid\
        \ JWT token\n- 400 BAD REQUEST: Invalid filter parameters or pagination\n\
        - 500 INTERNAL_SERVER_ERROR: Database error\n\nSecurity:\n- JWT token required:\
        \ All requests authenticated hone chahiye\n- Clinic isolation: Clinic-specific\
        \ exercises visible + global exercises\n- No role restriction: All authenticated\
        \ users can list\n\nFiltering Examples:\n- Body part filtering: GET /exercises?body_part=Shoulder\n\
        - Free exercises only: GET /exercises?is_free=true\n- Search: GET /exercises?search=rotation\n\
        - Combined: GET /exercises?body_part=Shoulder&is_free=true&search=rotation\n\
        \nPagination:\n- Server-side pagination: Large exercise libraries efficiently\
        \ handle\n- Total count: Frontend pagination UI ke liye\n- Page size limit:\
        \ DOS attack prevention\n\nGlobal Exercises:\n- clinic_id = null hote hain\n\
        - Sab clinics ke users ko visible\n- Shared resource across platform\n\nUsage\
        \ Example:\ncurl -X GET \"http://localhost:8000/api/v1/exercises?body_part=Shoulder&is_free=true&page=1\"\
        \ \\\n  -H \"Authorization: Bearer eyJhbGc...\""
      operationId: list_exercises_api_v1_exercises_get
      parameters:
      - name: body_part
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          description: Filter by body part
          title: Body Part
        description: Filter by body part
      - name: is_free
        in: query
        required: false
        schema:
          anyOf:
          - type: boolean
          - type: 'null'
          description: Filter by free/paid status
          title: Is Free
        description: Filter by free/paid status
      - name: search
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          description: Search in exercise title
          title: Search
        description: Search in exercise title
      - name: page
        in: query
        required: false
        schema:
          type: integer
          minimum: 1
          description: Page number
          default: 1
          title: Page
        description: Page number
      - name: page_size
        in: query
        required: false
        schema:
          type: integer
          maximum: 100
          minimum: 1
          description: Items per page
          default: 20
          title: Page Size
        description: Items per page
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResponseEnvelope_List_ExerciseRead__'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    post:
      tags:
      - Exercises
      - Exercises
      summary: Create Exercise
      description: "Endpoint ka purpose: New exercise create karna clinic mein (admin\
        \ only)\nYeh endpoint exercise creation handle karta hai.\n\nHTTP Method:\
        \ POST\nURL: /api/v1/exercises\nStatus Code: 201 CREATED (new exercise created)\n\
        \nRequest Body (ExerciseCreate):\n{\n    \"title\": \"Shoulder Rotation\"\
        ,                      # Required\n    \"description\": \"Step by step instructions...\"\
        ,     # Optional\n    \"body_part\": \"Shoulder\",                       \
        \   # Optional\n    \"is_free\": true,                                   #\
        \ Optional (default: false)\n    \"video_url\": \"https://cdn.example.com/video.mp4\"\
        \  # Optional\n}\n\nRequest Headers:\nAuthorization: Bearer {jwt_token}\n\
        Content-Type: application/json\n\nResponse (ResponseEnvelope[ExerciseRead]):\n\
        {\n    \"status\": \"success\",\n    \"data\": {\n        \"id\": \"new-uuid\"\
        ,\n        \"clinic_id\": \"clinic-uuid\",\n        \"title\": \"Shoulder\
        \ Rotation\",\n        \"description\": \"Step by step instructions...\",\n\
        \        \"body_part\": \"Shoulder\",\n        \"is_free\": true,\n      \
        \  \"video_url\": \"https://cdn.example.com/video.mp4\",\n        \"created_at\"\
        : \"2024-01-01T00:00:00Z\",\n        \"updated_at\": \"2024-01-01T00:00:00Z\"\
        \n    }\n}\n\nError Responses:\n- 401 UNAUTHORIZED: Missing or invalid JWT\
        \ token\n- 403 FORBIDDEN: User role is not admin\n- 400 BAD REQUEST: Invalid\
        \ request body or validation failure\n- 500 INTERNAL_SERVER_ERROR: Database\
        \ error\n\nSecurity:\n- JWT token required: All requests authenticated\n-\
        \ Admin only: Only admin role create kar sakte hain (no physios, patients)\n\
        - Clinic isolation: Exercise automatically current clinic se associated\n\
        - Input validation: Pydantic schema se strict validation\n\nAuthorization:\n\
        - Admin: Can create exercises\n- Physio: Cannot create (read-only access)\n\
        - Patient: Cannot create (no access)\n\nBusiness Logic:\n1. Authorization\
        \ check - admin role required\n2. Request data validation (Pydantic schema)\n\
        3. Exercise database record create\n4. Created exercise return\n\nVideo URL\
        \ Recommendations:\n- Use CDN URLs (AWS S3, Azure Blob, Cloudinary)\n- Support\
        \ multiple formats (mp4, webm, etc.)\n- Compression for fast loading\n\nUsage\
        \ Example:\ncurl -X POST http://localhost:8000/api/v1/exercises \\\n  -H \"\
        Authorization: Bearer eyJhbGc...\" \\\n  -H \"Content-Type: application/json\"\
        \ \\\n  -d '{\n    \"title\": \"Shoulder Rotation\",\n    \"description\"\
        : \"Rotate your shoulder in circular motion clockwise and counterclockwise.\"\
        ,\n    \"body_part\": \"Shoulder\",\n    \"is_free\": true,\n    \"video_url\"\
        : \"https://cdn.example.com/exercise-shoulder-rotation.mp4\"\n  }'"
      operationId: create_exercise_api_v1_exercises_post
      security:
      - OAuth2PasswordBearer: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ExerciseCreate'
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResponseEnvelope_ExerciseRead_'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/exercises/by-condition:
    get:
      tags:
      - Exercises
      - Exercises
      summary: Get Exercises By Condition
      description: 'Endpoint to filter exercises by condition name.

        Maps known conditions to exercise titles, or searches title/description.'
      operationId: get_exercises_by_condition_api_v1_exercises_by_condition_get
      parameters:
      - name: condition
        in: query
        required: true
        schema:
          type: string
          description: Condition name to filter by
          title: Condition
        description: Condition name to filter by
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResponseEnvelope_List_ExerciseRead__'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/exercises/{id}:
    get:
      tags:
      - Exercises
      - Exercises
      summary: Get Exercise
      description: "Endpoint ka purpose: Specific exercise ke details fetch karna\n\
        Yeh endpoint exercise details retrieve karta hai by ID.\n\nHTTP Method: GET\n\
        URL: /api/v1/exercises/{id}\nStatus Code: 200 OK (found)\n\nPath Parameters:\n\
        - id (str): Exercise UUID\n\nRequest Headers:\nAuthorization: Bearer {jwt_token}\n\
        \nResponse (ResponseEnvelope[ExerciseRead]):\n{\n    \"status\": \"success\"\
        ,\n    \"data\": {\n        \"id\": \"exercise-uuid\",\n        \"clinic_id\"\
        : \"clinic-uuid or null\",\n        \"title\": \"Shoulder Rotation\",\n  \
        \      \"description\": \"Detailed description\",\n        \"body_part\":\
        \ \"Shoulder\",\n        \"is_free\": true,\n        \"video_url\": \"https://cdn.example.com/exercise.mp4\"\
        ,\n        \"created_at\": \"2024-01-01T00:00:00Z\",\n        \"updated_at\"\
        : \"2024-01-01T00:00:00Z\"\n    }\n}\n\nError Responses:\n- 401 UNAUTHORIZED:\
        \ Missing or invalid JWT token\n- 404 NOT_FOUND: Exercise not found or not\
        \ accessible to user's clinic\n- 400 BAD REQUEST: Invalid exercise ID format\n\
        - 500 INTERNAL_SERVER_ERROR: Database error\n\nSecurity:\n- JWT token required:\
        \ All requests authenticated\n- Clinic + Global access: User's clinic exercises\
        \ + global exercises visible\n- No role restriction: All authenticated users\
        \ can view\n\nAccessible Exercises:\n- Clinic-specific exercises: clinic_id\
        \ matches user's clinic\n- Global exercises: clinic_id = null (shared across\
        \ all clinics)\n- Other clinics' exercises: Not accessible (404)\n\nBusiness\
        \ Logic:\n1. Exercise fetch by ID\n2. Clinic verification (clinic-specific\
        \ ya global)\n3. Exercise found = return data, Not found = 404\n\nUsage Example:\n\
        curl -X GET http://localhost:8000/api/v1/exercises/12345678-1234-5678-1234-567812345678\
        \ \\\n  -H \"Authorization: Bearer eyJhbGc...\"\n\nUse Cases:\n- Get exercise\
        \ details for patient view\n- Fetch video URL for streaming\n- Load exercise\
        \ instructions\n- Retrieve exercise metadata for prescriptions"
      operationId: get_exercise_api_v1_exercises__id__get
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          title: Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResponseEnvelope_ExerciseRead_'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    put:
      tags:
      - Exercises
      - Exercises
      summary: Update Exercise
      description: Update an existing exercise.
      operationId: update_exercise_api_v1_exercises__id__put
      security:
      - OAuth2PasswordBearer: []
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          title: Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ExerciseUpdate'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResponseEnvelope_ExerciseRead_'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    delete:
      tags:
      - Exercises
      - Exercises
      summary: Delete Exercise
      description: Delete an exercise.
      operationId: delete_exercise_api_v1_exercises__id__delete
      security:
      - OAuth2PasswordBearer: []
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          title: Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResponseEnvelope_dict_'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
...
    ExerciseCreate:
      properties:
        title:
          type: string
          title: Title
        description:
          anyOf:
          - type: string
          - type: 'null'
          title: Description
        body_part:
          anyOf:
          - type: string
          - type: 'null'
          title: Body Part
        is_free:
          type: boolean
          title: Is Free
          default: false
        video_url:
          anyOf:
          - type: string
          - type: 'null'
          title: Video Url
      type: object
      required:
      - title
      title: ExerciseCreate
    ExerciseRead:
      properties:
        title:
          type: string
          title: Title
        description:
          anyOf:
          - type: string
          - type: 'null'
          title: Description
        body_part:
          anyOf:
          - type: string
          - type: 'null'
          title: Body Part
        is_free:
          type: boolean
          title: Is Free
          default: false
        video_url:
          anyOf:
          - type: string
          - type: 'null'
          title: Video Url
        id:
          type: string
          format: uuid
          title: Id
        clinic_id:
          anyOf:
          - type: string
            format: uuid
          - type: 'null'
          title: Clinic Id
        created_at:
          type: string
          format: date-time
          title: Created At
        updated_at:
          type: string
          format: date-time
          title: Updated At
      type: object
      required:
      - title
      - id
      - clinic_id
      - created_at
      - updated_at
      title: ExerciseRead
    ExerciseUpdate:
      properties:
        title:
          anyOf:
          - type: string
          - type: 'null'
          title: Title
        description:
          anyOf:
          - type: string
          - type: 'null'
          title: Description
        body_part:
          anyOf:
          - type: string
          - type: 'null'
          title: Body Part
        is_free:
          anyOf:
          - type: boolean
          - type: 'null'
          title: Is Free
        video_url:
          anyOf:
          - type: string
          - type: 'null'
          title: Video Url
      type: object
      title: ExerciseUpdate
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    MetaPagination:
      properties:
        total:
          type: integer
          title: Total
        page:
          type: integer
          title: Page
        page_size:
          type: integer
          title: Page Size
      type: object
      required:
      - total
      - page
      - page_size
      title: MetaPagination
...
    ResponseEnvelope_ExerciseRead_:
      properties:
        data:
          anyOf:
          - $ref: '#/components/schemas/ExerciseRead'
          - type: 'null'
        meta:
          anyOf:
          - $ref: '#/components/schemas/MetaPagination'
          - $ref: '#/components/schemas/MetaError'
          - type: object
          - type: 'null'
          title: Meta
      type: object
      title: ResponseEnvelope[ExerciseRead]
    ResponseEnvelope_List_ExerciseRead__:
      properties:
        data:
          anyOf:
          - items:
              $ref: '#/components/schemas/ExerciseRead'
            type: array
          - type: 'null'
          title: Data
        meta:
          anyOf:
          - $ref: '#/components/schemas/MetaPagination'
          - $ref: '#/components/schemas/MetaError'
          - type: object
          - type: 'null'
          title: Meta
      type: object
      title: ResponseEnvelope[List[ExerciseRead]]
...
    ResponseEnvelope_dict_:
      properties:
        data:
          anyOf:
          - type: object
          - type: 'null'
          title: Data
        meta:
          anyOf:
          - $ref: '#/components/schemas/MetaPagination'
          - $ref: '#/components/schemas/MetaError'
          - type: object
          - type: 'null'
          title: Meta
      type: object
      title: ResponseEnvelope[dict]
...
  securitySchemes:
    OAuth2PasswordBearer:
      type: oauth2
      flows:
        password:
          scopes: {}
          tokenUrl: /api/v1/auth/login